Modify gdalinfo json to contain final asset path. https://github.com/Open-EO/openeo-geotrellis-extensions/issues/352

+++ geotrellis-s3-extensions/src/main/scala/org/openeo/geotrelliss3/S3Utils.scala
@@ -9,20 +9,23 @@ object S3Utils {
 
   val logger = LoggerFactory.getLogger(S3Utils.getClass)
 
-  def deleteSubFolder(client: S3Client, bucketName: String, subfolder: String) = {
+  def deleteSubFolder(client: S3Client, bucketName: String, subfolder: String): Unit = {
     val listObjectsRequest = ListObjectsRequest.builder
       .bucket(bucketName)
       .prefix(subfolder)
       .build
     val listObjectsResponse = client.listObjects(listObjectsRequest)
     val keys = listObjectsResponse.contents.asScala.map(_.key)
+    if (keys.isEmpty) {
+      logger.info(s"No objects to delete in $bucketName/$subfolder")
+      // Avoid S3Exception: The XML you provided was not well-formed or did not validate against our published schema
+      return
+    }
     val deleteObjectsRequest = DeleteObjectsRequest.builder
       .bucket(bucketName)
       .delete(Delete.builder.objects(keys.map(key => ObjectIdentifier.builder.key(key).build).asJavaCollection).build)
       .build
-    logger.info("Deleting objects from S3.")
-    logger.info(s"Bucket: $bucketName, Subfolder: $subfolder")
-    logger.info(s"Objects: ${keys.mkString(", ")}")
+    logger.info(s"Deleting objects from $bucketName/$subfolder: ${keys.mkString(", ")}")
     client.deleteObjects(deleteObjectsRequest)
   }
